Now wokring. LCD off. Comms every 200ms. NRF24L01 lib is changed.
