+++ platformio.ini
@@ -1,8 +1,8 @@
 [platformio]
 ; default_envs = esp32dev.debug
-default_envs = esp32dev.release
-; default_envs = esp32dev.test_comms
-; default_envs = esp32dev.test_comms_and_display
+; default_envs = esp32dev.release
+; default_envs = esp32dev.comms_only
+default_envs = esp32dev.comms_and_display
 
 [env]
 platform = espressif32
@@ -66,12 +66,12 @@ build_flags =
   -D PRINT_RETRIES_STATS
 ;-------------------------------------------------------
 
-[env:esp32dev.test_comms]
-src_filter = ${dir.test_comms}
+[env:esp32dev.comms_only]
+src_filter = ${dir.comms_only}
 ;-------------------------------------------------------
 
-[env:esp32dev.test_comms_and_display]
-src_filter = ${dir.test_comms_and_display}
+[env:esp32dev.comms_and_display]
+src_filter = ${dir.comms_and_display}
 
 lib_deps = 
   https://github.com/lewisxhe/Button2.git
@@ -88,6 +88,6 @@ lib_deps =
 
 ;-------------------------------------------------------
 [dir]
-main = +<main/> -<test_comms/> -<comms_and_display/>
-test_comms = +<test_comms/>
-test_comms_and_display = +<comms_and_display/>
+main = +<main/> -<comms_only/> -<comms_and_display/>
+comms_only = +<comms_only/>
+comms_and_display = +<comms_and_display/>
